Split code generators out of the main annotation processor class.
Moved validators to 'validators' package and generators to 'generators' package.
