+++ robolectric-processor/src/main/java/org/robolectric/annotation/processing/generator/Generator.java
@@ -1,0 +1,7 @@
+package org.robolectric.annotation.processing.generator;
+
+public abstract class Generator {
+  protected static final String GEN_CLASS = "Shadows";
+
+  public abstract void generate(String shadowPackage);
+}
