+++ robolectric-processor/src/test/java/org/robolectric/annotation/processing/validator/RealObjectValidatorTest.java
@@ -1,14 +1,15 @@
-package org.robolectric.annotation.processing;
+package org.robolectric.annotation.processing.validator;
 
 import static com.google.testing.compile.JavaFileObjects.forResource;
 import static com.google.testing.compile.JavaSourcesSubjectFactory.javaSources;
 import static org.truth0.Truth.ASSERT;
-import static org.robolectric.annotation.processing.Utils.SHADOW_EXTRACTOR_SOURCE;
-import static org.robolectric.annotation.processing.SingleClassSubject.singleClass;
+import static org.robolectric.annotation.processing.validator.Utils.SHADOW_EXTRACTOR_SOURCE;
+import static org.robolectric.annotation.processing.validator.SingleClassSubject.singleClass;
 
 import org.junit.Test;
 
 import com.google.common.collect.ImmutableList;
+import org.robolectric.annotation.processing.RoboProcessor;
 
 public class RealObjectValidatorTest {
   @Test
