+++ robolectric-processor/src/test/java/org/robolectric/annotation/processing/validator/ResetterValidatorTest.java
@@ -1,7 +1,7 @@
-package org.robolectric.annotation.processing;
+package org.robolectric.annotation.processing.validator;
 
 import static org.truth0.Truth.ASSERT;
-import static org.robolectric.annotation.processing.SingleClassSubject.singleClass;
+import static org.robolectric.annotation.processing.validator.SingleClassSubject.singleClass;
 
 import org.junit.Test;
